- Fixed a large number of type conversion warnings  - Switch file_reporter to use boost::format instead of fprintf

+++ src/graphlab/factors/binary_factor.cpp
@@ -23,8 +23,8 @@ std::ostream& operator<<(std::ostream& out,
       << fact.arity1() << "}, " 
       << ", v_ " << fact.var2() << " in {1..." 
       << fact.arity2() << "})" << std::endl;
-  for(size_t i = 0; i < fact.arity1(); ++i) {
-    for(size_t j = 0; j < fact.arity2(); ++j) {
+  for(uint16_t i = 0; i < fact.arity1(); ++i) {
+    for(uint16_t j = 0; j < fact.arity2(); ++j) {
       out << fact.logP(i,j) << " ";
     }
     out << std::endl;
